LayoutTests/imported/w3c/web-platform-tests/css/css-values/line-break-ch-unit-expected.txt


PASS 3ch should fit
PASS 4ch should fit
PASS 5ch should fit
PASS 6ch should fit
PASS 7ch should fit
PASS 8ch should fit
PASS 9ch should fit
PASS 10ch should fit
PASS 11ch should fit
PASS 12ch should fit
PASS 13ch should fit
PASS 14ch should fit
PASS 15ch should fit
PASS 16ch should fit
PASS 17ch should fit
PASS 18ch should fit
PASS 19ch should fit
PASS 20ch should fit
PASS 21ch should fit
PASS 22ch should fit
PASS 23ch should fit
PASS 24ch should fit
PASS 25ch should fit
PASS 26ch should fit
PASS 27ch should fit
PASS 28ch should fit
PASS 29ch should fit
PASS 30ch should fit
PASS 31ch should fit
PASS 32ch should fit
PASS 33ch should fit
PASS 34ch should fit
PASS 35ch should fit
PASS 36ch should fit
PASS 37ch should fit
PASS 38ch should fit
PASS 39ch should fit
PASS 40ch should fit
PASS 41ch should fit
PASS 42ch should fit
PASS 43ch should fit
PASS 44ch should fit
PASS 45ch should fit
PASS 46ch should fit
PASS 47ch should fit
PASS 48ch should fit
PASS 49ch should fit
PASS 50ch should fit
PASS 51ch should fit
PASS 52ch should fit
PASS 53ch should fit
PASS 54ch should fit
PASS 55ch should fit
PASS 56ch should fit
PASS 57ch should fit
PASS 58ch should fit
PASS 59ch should fit
PASS 60ch should fit
PASS 61ch should fit
PASS 62ch should fit
PASS 63ch should fit
PASS 64ch should fit
PASS 65ch should fit
PASS 66ch should fit
PASS 67ch should fit
PASS 68ch should fit
PASS 69ch should fit
PASS 70ch should fit
PASS 71ch should fit
PASS 72ch should fit
PASS 73ch should fit
PASS 74ch should fit
PASS 75ch should fit
PASS 76ch should fit
PASS 77ch should fit
PASS 78ch should fit
PASS 79ch should fit
PASS 80ch should fit
PASS 81ch should fit
PASS 82ch should fit
PASS 83ch should fit
PASS 84ch should fit
PASS 85ch should fit
PASS 86ch should fit
PASS 87ch should fit
PASS 88ch should fit
PASS 89ch should fit
PASS 90ch should fit
PASS 91ch should fit
PASS 92ch should fit
PASS 93ch should fit
PASS 94ch should fit
PASS 95ch should fit
PASS 96ch should fit
PASS 97ch should fit
PASS 98ch should fit
PASS 99ch should fit
PASS 3ch should wrap
PASS 4ch should wrap
PASS 5ch should wrap
PASS 6ch should wrap
PASS 7ch should wrap
PASS 8ch should wrap
PASS 9ch should wrap
PASS 10ch should wrap
PASS 11ch should wrap
PASS 12ch should wrap
PASS 13ch should wrap
PASS 14ch should wrap
PASS 15ch should wrap
PASS 16ch should wrap
PASS 17ch should wrap
PASS 18ch should wrap
PASS 19ch should wrap
PASS 20ch should wrap
PASS 21ch should wrap
PASS 22ch should wrap
PASS 23ch should wrap
PASS 24ch should wrap
PASS 25ch should wrap
PASS 26ch should wrap
PASS 27ch should wrap
PASS 28ch should wrap
PASS 29ch should wrap
PASS 30ch should wrap
PASS 31ch should wrap
PASS 32ch should wrap
PASS 33ch should wrap
PASS 34ch should wrap
PASS 35ch should wrap
PASS 36ch should wrap
PASS 37ch should wrap
PASS 38ch should wrap
PASS 39ch should wrap
PASS 40ch should wrap
PASS 41ch should wrap
PASS 42ch should wrap
PASS 43ch should wrap
PASS 44ch should wrap
PASS 45ch should wrap
PASS 46ch should wrap
PASS 47ch should wrap
PASS 48ch should wrap
PASS 49ch should wrap
PASS 50ch should wrap
PASS 51ch should wrap
PASS 52ch should wrap
PASS 53ch should wrap
PASS 54ch should wrap
PASS 55ch should wrap
PASS 56ch should wrap
PASS 57ch should wrap
PASS 58ch should wrap
PASS 59ch should wrap
PASS 60ch should wrap
PASS 61ch should wrap
PASS 62ch should wrap
PASS 63ch should wrap
PASS 64ch should wrap
PASS 65ch should wrap
PASS 66ch should wrap
PASS 67ch should wrap
PASS 68ch should wrap
PASS 69ch should wrap
PASS 70ch should wrap
PASS 71ch should wrap
PASS 72ch should wrap
PASS 73ch should wrap
PASS 74ch should wrap
PASS 75ch should wrap
PASS 76ch should wrap
PASS 77ch should wrap
PASS 78ch should wrap
PASS 79ch should wrap
PASS 80ch should wrap
PASS 81ch should wrap
PASS 82ch should wrap
PASS 83ch should wrap
PASS 84ch should wrap
PASS 85ch should wrap
PASS 86ch should wrap
PASS 87ch should wrap
PASS 88ch should wrap
PASS 89ch should wrap
PASS 90ch should wrap
PASS 91ch should wrap
PASS 92ch should wrap
PASS 93ch should wrap
PASS 94ch should wrap
PASS 95ch should wrap
PASS 96ch should wrap
PASS 97ch should wrap
PASS 98ch should wrap
PASS 99ch should wrap
0 0
0 00
0 000
0 0000
0 00000
0 000000
0 0000000
0 00000000
0 000000000
0 0000000000
0 00000000000
0 000000000000
0 0000000000000
0 00000000000000
0 000000000000000
0 0000000000000000
0 00000000000000000
0 000000000000000000
0 0000000000000000000
0 00000000000000000000
0 000000000000000000000
0 0000000000000000000000
0 00000000000000000000000
0 000000000000000000000000
0 0000000000000000000000000
0 00000000000000000000000000
0 000000000000000000000000000
0 0000000000000000000000000000
0 00000000000000000000000000000
0 000000000000000000000000000000
0 0000000000000000000000000000000
0 00000000000000000000000000000000
0 000000000000000000000000000000000
0 0000000000000000000000000000000000
0 00000000000000000000000000000000000
0 000000000000000000000000000000000000
0 0000000000000000000000000000000000000
0 00000000000000000000000000000000000000
0 000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00
0 000
0 0000
0 00000
0 000000
0 0000000
0 00000000
0 000000000
0 0000000000
0 00000000000
0 000000000000
0 0000000000000
0 00000000000000
0 000000000000000
0 0000000000000000
0 00000000000000000
0 000000000000000000
0 0000000000000000000
0 00000000000000000000
0 000000000000000000000
0 0000000000000000000000
0 00000000000000000000000
0 000000000000000000000000
0 0000000000000000000000000
0 00000000000000000000000000
0 000000000000000000000000000
0 0000000000000000000000000000
0 00000000000000000000000000000
0 000000000000000000000000000000
0 0000000000000000000000000000000
0 00000000000000000000000000000000
0 000000000000000000000000000000000
0 0000000000000000000000000000000000
0 00000000000000000000000000000000000
0 000000000000000000000000000000000000
0 0000000000000000000000000000000000000
0 00000000000000000000000000000000000000
0 000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
